+++ src/SobchakMain.xaml.cs
@@ -1,0 +1,22 @@
+﻿/* PROJECT: Sobchak (https://github.com/aprettycoolprogram/Sobchak)
+ *    FILE: Sobchak.SobchakMain.xaml.cs
+ * UPDATED: 1-20-2021-8:57 AM
+ * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
+ *          Copyright 2020 A Pretty Cool Program All rights reserved
+ */
+
+using System.Windows;
+
+namespace Sobchak
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+    }
+}