+++ src/Resources/Doc/changelog.md
@@ -1,4 +1,15 @@
 ﻿# Sobchak: Changelog
 
-#### v0.0
-`INFO` Initial commit
+#### 0.1
+> Project framework updates
+
+#### v0.1.21020.1358 (2021-01-20)
+* `ADDED` du.proj
+* `UPDATED` Documentation
+* `UPDATED` sobchak.licenseheader
+
+## Version 0.0 - January 19, 2021
+> Initial commit, no working code.
+
+#### v0.0.0.0 (2021-01-19)
+* `ADDED` Initial project framework